Clone the child node when the decorator is cloned (#707)
While investigating Expensify/App#65268 I noticed that the node with display: contents set can be cloned without cloning its children. Currently, the decorator assumes that when it's mutable, its child is mutable as well, which is clearly wrong in some cases.

This PR adds a step cloning the child node when the decorator is cloned, which makes the above assumption correct.

Note that this change does not fix the linked issue.